Extract MathHelper module
[ci skip]
